services: Fix double-free error on RemoteFillService
Symptom:
RemoteFillService was crashed due to IllegalArgumentException
"Service not registered:" at onServiceConnected.

Root cause:
RemoteFillService#onServiceConnected tries to unbind the connection
if mDestroyed is flagged or mBinding is not flagged. It always fails
with IllegalArgumentException.
Both mDestroyed and !mBinding mean the connection was unbound.
You can't unbind the unbound connection. It's not allowed.

Change-Id: If5481468ddac7be41accad63e9d5382bc6c029fd
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>
Signed-off-by: Adesh15 <adesikha15@gmail.com>